+++ src/main/java/service/IAuthService.java
@@ -1,0 +1,9 @@
+package service;
+
+import dto.AuthRequest;
+
+import java.io.IOException;
+
+public interface IAuthService {
+    String getAuthToken(AuthRequest request) throws IOException;
+}
